Ajout de server Controller

+++ Certification_Backend/src/main/java/com/kwiky/backend/service/ServerDirectory.java
@@ -7,6 +7,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.kwiky.backend.dao.ServerRepository;
+import com.kwiky.backend.model.Message;
 import com.kwiky.backend.model.Server;
 
 @Service
@@ -15,8 +16,8 @@ public class ServerDirectory {
 	ServerRepository sr;
 	
 		// post server
-		public void addServer(Server newServer) {
-			sr.save(newServer);
+		public Server addServer(Server newServer) {
+			return sr.save(newServer);
 		}
 
 		// Get All Canal
@@ -30,12 +31,20 @@ public class ServerDirectory {
 		}
 
 		// delete server by id
-		public void deleteServer(Long id) {
-			sr.deleteById(id);
+		public boolean deleteServer(Long id) {
+			boolean b = sr.existsById(id);
+			if(b)
+				sr.deleteById(id);
+			return b;
 		}
 
-		// Put server by id
-		public void putServer(Server serverToUpdate, Long id) {
-			sr.save(serverToUpdate);
+		/**Renvoie true si l'id existait*/
+		public boolean updateServer(Server server)
+		{
+			boolean b = sr.existsById(server.getId());
+			if(b)
+				sr.save(server);
+			
+			return b;
 		}
 }
